refactor api handlers
