+++ api/handler_time.go
@@ -1,4 +1,4 @@
-package handler
+package api
 
 import (
 	"strconv"
@@ -15,7 +15,7 @@ import (
 //	@Success		200	{object}	resource.ResGetTime
 //	@Failure		500	{object}	resource.ResErr
 //	@Router			/api/ledger/time [get]
-func Time(c *gin.Context) {
+func (s *Server) Time(c *gin.Context) {
 	now := time.Now()
 	timestamp := strconv.FormatInt(now.Unix(), 10)
 	date := now.Format("2006-01-02 15:04")
